requirements.txt
asgiref==3.5.2
autopep8==2.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.6
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.6
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-bootstrap-icons==0.8.2
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
drf-yasg==1.21.3
filelock==3.8.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
pipenv==2022.11.25
platformdirs==2.5.4
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.11
virtualenv==20.16.7
virtualenv-clone==0.5.7
whitenoise==6.2.0
